+++ CHANGELOG.md
@@ -26,7 +26,7 @@ Changes
     ```
 * You can now add or remove multiple permissions at the same time with `!op` and `!deop`
     + Example: `!op <user id> general:kill general:reload`
-* You can now give the bot an avatar with `!avatar <url or relative path>` or by setting it in your `config.json`
+* You can now give the bot an avatar with `!avatar <url or relative path>` or by setting it in your `config.json`. Setting it to `null` will remove the avatar.
     + Example:
     ```json
     "credentials": {